requirements.txt
# ===== 核心依赖 =====
# Binance API客户端 (自动包含 requests, pycryptodome, websocket-client)
binance-futures-connector>=4.1.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 技术指标计算
TA-Lib>=0.4.0

# 环境变量管理
python-dotenv>=1.0.0

# ===== 说明 =====
# 以下依赖会被 binance-futures-connector 自动安装，无需单独指定：
# - requests>=2.25.0 (HTTP请求)
# - pycryptodome>=3.15.0 (加密)
# - websocket-client>=1.5.0 (WebSocket客户端)
